refactor(code-input): extract fullscreen into its own module; fix tooltip
- Move all fullscreen logic, styled components and the toggle button out of
  CodeInput.tsx into a new CodeInputFullscreen.tsx (exposed as a FullscreenEditor
  wrapper), keeping CodeInput.tsx focused on the field itself.
- Place the toggle tooltip below the button in both states for consistency.
- Constrain the tooltip to the overlay so a surrounding modal's boundary no
  longer pushes it away from the button when expanded.
